+++ src/frame_compare/vs/__init__.py
@@ -7,8 +7,9 @@ from frame_compare.vs.env import (
     require_plugin,
 )
 from frame_compare.vs.loader import DefaultVSLoader, VSLoader
+from frame_compare.vs.props import get_color_props, is_hdr
 from frame_compare.vs.source import apply_trim, load_source
-from frame_compare.vs.types import HDRMetadata, SourceInfo, TonemapSettings
+from frame_compare.vs.types import ColorProps, HDRMetadata, SourceInfo, TonemapSettings
 
 __all__ = [
     "VSLoader",
@@ -16,10 +17,13 @@ __all__ = [
     "SourceInfo",
     "HDRMetadata",
     "TonemapSettings",
+    "ColorProps",
     "is_vapoursynth_available",
     "ensure_vs_environment",
     "detect_plugins",
     "require_plugin",
     "load_source",
     "apply_trim",
+    "get_color_props",
+    "is_hdr",
 ]